move screenshots to screenshots_name
